+++ changelog.md
@@ -55,6 +55,7 @@ This version is not yet released. If you are reading this on the website, then t
 - When hovering a user-defined function's name, the language server will now show whether it is compatible with [`un °`](https://uiua.org/docs/un) and [`under ⍜`](https://uiua.org/docs/under)
 - Optimize the pattern for adjacency: `≡/F◫2`
   - It it as much as 800x faster in some cases
+- Square matrices are now transposed in-place
 - Add completions to the language server
 ### Website
 - Add [Tutorial Introduction](https://uiua.org/tutorial/introduction)